requirements.txt
pydantic==2.11.7
requests==2.31.0
playwright==1.54.0
pillow==11.1.0
beautifulsoup4==4.13.5
colorlog==6.9.0
